Add filter for all files in all project directories
Behaves like the custom directory filters, but with preset directories,
updating whenever a project is loaded or closed.
Loading or closing a project will trigger a refresh of the file cache,
otherwise the filter refreshes with the refresh interval that the other
custom directory filters use, or on manual request.

Task-number: QTCREATORBUG-19122
Change-Id: Id4f65b5137284e63412390672fa8449030c9fed4
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>