Fix MsvcToolChain::fromMap
Previous versions did not store supported abis so detect these if they
cannot be loaded. Having supported Abis is now part of the toolchain's
validity, so also check these in the condition.

Additionally, only valid toolchains should be added to the global list
of Msvc toolchains, so another check was added.

Change-Id: I5e4b42e947210611c16a002a2b50c44aed2b5b22
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>