Fix qdoc call after it moved to LLVM/Clang
For some unknown reason qdoc does not add the include paths for
Qt modules, so we need to do that manually.
Also, if we do that, we need to pass it the default include paths, so
standard C++ headers are found.

Qt includes and standard headers are found now, but

Framework style includes on macOS still fail ("QtCore/QString"), which are
used in Qt headers as well. Looks like the framework path that is passed
with -F to qdoc is not passed on correctly.

Task-number: QTCREATORBUG-22451
Change-Id: I04b2c75ecdeb1f0e70ba9adfea039f0ff16ec96b
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>