CMake build: Read QTC_BUILD_PLUGIN_${NAME} environment as default
for BUILD_PLUGIN_${NAME}. This is only used as the default for the CMake
cache, so only if you haven't run CMake before.
The disadvantage is that just changing the environment variable after
the fact will not change anything.
The advantage is that the CMake cache keeps reflecting the actual state,
which is great for ccmake, the UI in Qt Creator, and other configuration
tools.

Change-Id: I25448270c29afa31d240072f58f981b8776597f2
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>