ProjectTree: Fix crash when project has no rootProjectNode
Fix a crash that is triggered by a project returning to a state
where it has no rootProjectNode. This can happen when parsing fails
and Creator should fall back to displaying the project name and
its main project file as it does before any parsing had been done.

Unfortunately the hasNode function returned false in this case, so
the project model was never updated and the removed project nodes
stuck around, triggering a crash.

Change-Id: I7616e576773dc52fb6fdff39b9f0a7c7729eac71
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>