Fix disabled close button in designer editor tool bar
It was never actually updating when switching the editor.
An update was just accidentally triggered when the designer plugin
updated the editor XML when switching modes, which triggered
an intermediate document change signal.

Get rid of the separate code paths for setting the current editor
for "standalone" editor tool bars and the tool bars in editor views,
which implicitly corrects the update behavior in the former case.

Fixes: QTCREATORBUG-22553
Change-Id: Ieb9f4b53600e1e1b66695ec86164628025df73f8
Reviewed-by: David Schulz <david.schulz@qt.io>