AutoTest: Limit condition for test execution to Linux
gcc and clang are handling builtin include paths differently.
On Linux the default should be gcc - which works fine even
if boost is installed in /usr/local.
On macOS clang is default which does have the correct builtin
include paths only for Qbs but we are testing on qmake as well.
So, expect the BOOST_INCLUDE_DIR set on macOS if we want to
have this test executed.

Change-Id: Ib9ad697b59bfad5413ad984c9b969ec8b174ae9a
Reviewed-by: David Schulz <david.schulz@qt.io>