ProjectExplorer: Move setDisplayName code out of *ToolChain::create()
This sets the display name in the constructor now. In quite a few
places this will be directly overridden afterwards (either in
derived c'tor or in detectToolChains(), but this will help with
further changes:

The idea is to only leave  { return new ... } in the create()
implementation such the method can be de-virtualized as the base
ToolChainFactory::create is enough.

Change-Id: I1b15c015b6f3e2bbff682bee3fac626ed1bde312
Reviewed-by: David Schulz <david.schulz@qt.io>