ensure the 'Auto' key is set while registering post mortem debugger
The 'Auto' indicates that the the registered post mortem debugger is
directly started after an unhandled user exception. If the value is
unset or set to '0' in rescent windows versions the debugger selection
dialog is not shown. Set this value to '1 ' to show the qtcdebugger
dialog, which also has an option to show the system dialog, if the user
does not want to debug the exception with Qt Creator.

Change-Id: I3160315060dbfb37bec5eaa677c4208900f574a4
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>