Escape tooltip text
The tooltip text may contain characters that need
to be escaped without 'Qt::mightBeRichText' categorizing
this as rich text.

This is needed e.g. to show the proper type when
hovering over a variable of type 'std::vector<int>'
(or some smart pointer) in edit view (no debugging
session).

Otherwise, just "std::vector" is shown since
'Qt::mightBeRichText' returns false for that string
and "<int>" is then probably just ignored as an invalid
tag when the tooltip is displayed.

Change-Id: I25ded2b21e4350a1036ab4b3d8f144383aee776d
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: hjk <hjk@qt.io>