QtOutputFormatter: Support relative file:// links
QUrl does not have good support for relative file paths, and it
seems there is little interest to change it: QTBUG-44921

Work around this by constructing the QUrl again on failure with
fromLocalFile(). In this case, the "file://" scheme has to be omitted.

This issue was detected because some compilers generate paths with
native Windows separators, like "..\main.cpp" for the __FILE__ macro.

Debugging the issue showed, that rather the relative path was the
problem, as the QUrl constructor does not support it.

Change-Id: I30074a3d11f9f43e05cd202f6d86c752ce184785
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>