Issues pane: Show the right number of issues on the pane button
While we do want to show contextual messages from the compiler, the
number displayed on the button should correspond to the actual number of
issues (i.e. errors or warnings), and should not include the notes
around them.
We used to do this correctly, but broke it in acd0d02e282f.

Fixes: QTCREATORBUG-18490
Change-Id: Ib6b36f24b4f6bfe53ed5b977f0f88586c4d95b16
Reviewed-by: hjk <hjk@qt.io>