Filter leading GDB messages during version detection
When using GDB through a script, additional messages can be printed
before the line reporting the version number, and some digits of those
messages could be interpreted as a version value, leading to wrong
version detection.

This patch searches for the "GNU gdb" tag first, so the correct
line is evaluated.

Change-Id: Ief64a056a20a963ad01c44941777361abcce00fc
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: hjk <hjk@qt.io>