ProjectExplorer: Auto-detect installed GNU tools for AVR on Windows
This patch implements the auto-detection for the installed
GCC embedded toolchains for AVR architecture provided
by "Microchip" (e.g. located in Atmel Studio):

https://www.microchip.com/mplab/avr-support/atmel-studio-7

The detection algorithm does search a toolchain path in a
Windows registry; this will work only if the Atmel Studio
was installed.

Tested on Windows 10 with Atmel Studio v6.2 installed.

Change-Id: Ibd2dd95daf7e9da272611fc32d679852a7f20952
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>