GenericProposalModel: Allow substring matches
So far, if we had these three functions:

  int getSomething() { return 44; }
  int get_something() { return 43; }
  int something() { return 42; }

then we did not get a completion for getSomething()
or get_something() if we just typed "some", as only
the prefix was taken into account.

This patch adds support for substring matches, and
adds these with lower priority to the proposal list.

Task-number: QTCREATORBUG-19170
Fixes: QTCREATORBUG-19918
Change-Id: Ifc5e2149e4b1fa995f1f8550efc84e7ff4fb1522
Reviewed-by: David Schulz <david.schulz@qt.io>