QtSupport: Fix handling of not yet saved *.ui files
Since

  commit fb3e18bd4ba7c08346290cf498f28bf82a7162d9
  Fix missing header guards in UI files

the uic input comes from a file (on disk) and thus the source contents
from the edited and not yet saved *.ui file (Designer mode) was ignored.

Revert the relevant part. This should not introduce a regression as

  commit 47312b04c0878e0c308794cde08e92c03d4be29e
  QtSupport: Fix header guard for ui files

basically superseded the first change, handling the proper inclusion of
the generated header file.

Fixes: QTCREATORBUG-22740
Change-Id: Iec6550817941eea99f823c377680b9a6fa5564e8
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>