BareMetal: Add STM8 architecture for SDCC toolchain
Previous implementation of SDCC toolchain based on an assumption,
that the default target architecture always is MCS51.

But, the SDCC compiler support multiple architectures:

* http://sdcc.sourceforge.net/

Since we have the new STM8 architecture in the Abi::Architecture
enumeration, it is makes sense to add support for this architecture
and to the SDCC compiler.

Right now, the SDCC compiler will be auto-detected as two instances:

* with mcs51 architecture
* with stm8 architecture

by analogy as it does for the MSVC compiler for multiple x86, x86_64
and so forth architectures.

In case the SDCC compiler is manually added, the user can choose
a desired architecture (mcs51 or stm8) from the combo-box.

Change-Id: I9b7d34f9064ea34c94dab7f1a7e3b64b190f7cb0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>