ProjectExplorer: Move *ConfigurationModel ownership to target
The models hold essentially duplicated information besides the actual
configuration lists that are already in the target and are only
used in a context where the target is known and accessible.

As the models keep track of all added or removed configurations
and can be fed configurations of the right types only, the need
to filter is gone.

Change-Id: Iec2c6879223893297c71e7632310ac526b704237
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>