Unification of desktop run configurations, step 3
Use a QtSupport::QtOutputFormatter unconditionally.

CMake previously only added one when there was a Qt version selected
in the kit at RunConfiguration construction time, which is an odd
moment for such a decision.

Arguably, the adding of output formatters should not be done
unconditionally either, but deferred to RunControl runtime
in principle. But that'd be an indpendent global change.

Change-Id: I8365ccf194a896d4cba5d9148cc0b21d83de2f3f
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>