ProjectExplorer: Persist run config build keys as separate entry
The id mangling is not needed by the upper layers anymore, time to
get rid of it in the .user files as well. For now, start storing
the build keys explicitly, use that directly when present, and fall
back to the old mechanism in cases we read settings without it.

Change-Id: I1fa467a1488828ffd400e7d67d6fcd8247c938d1
Reviewed-by: Christian Stenger <christian.stenger@qt.io>