ProjectExplorer: Do not disable run configs because of project parsing
After project parsing has ended, the run configurations are set up from
the information provided by the parse, and that information is valid
until the end of the next parse. There does not appear to be a reason to
prevent users from e.g. editing command line arguments or starting the
application just because the project is currently being parsed.

Task-number: QTCREATORBUG-22548
Change-Id: Ib5e77c747fc2483650bcab22e07881ee7d93500b
Reviewed-by: hjk <hjk@qt.io>