Tracing: Add hasMixedTypesInExpandedState property to TimelineModel
If the property is true, it disables the functionality to select the
type of row by clicking on the row label.
This is required by the CtfVisualizer plugin that shows more than one
type of events per row in the expanded state which would make the above
mentioned functionality meaningless.
This change doesn't change the current default behavior.

Change-Id: I20d42404a4cd3c9142bf0b59b57f4de083be07f3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>