QmlDesigner: Make AnimationCurveDialog a normal modal dialog
QDialog::open() opens the dialog as a modal dialog.
Since no dialog parent was set the dialog did not
block any parent windows.

In this patch we set Core::ICore::dialogParent as the parent
and use show() instead of open().
The dialog is not modal anymore and Qt Creator is the proper parent window.
Therefore Qt::WindowStaysOnTopHint is not required anymore.

Change-Id: Iea48a73e67de5e9d2568e71145d57d0ffd0df78b
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>